[Enhancement] Optimize hive table change mv refresh (#45118)

+++ fe/fe-core/src/main/java/com/starrocks/connector/hive/HiveMetadata.java
@@ -16,6 +16,7 @@ package com.starrocks.connector.hive;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.starrocks.catalog.Column;
 import com.starrocks.catalog.Database;
@@ -26,19 +27,28 @@ import com.starrocks.catalog.Table;
 import com.starrocks.common.AlreadyExistsException;
 import com.starrocks.common.DdlException;
 import com.starrocks.common.MetaNotFoundException;
+import com.starrocks.common.UserException;
+import com.starrocks.common.Version;
 import com.starrocks.connector.ConnectorMetadata;
 import com.starrocks.connector.HdfsEnvironment;
 import com.starrocks.connector.PartitionInfo;
+import com.starrocks.connector.PartitionUtil;
+import com.starrocks.connector.RemoteFileDesc;
 import com.starrocks.connector.RemoteFileInfo;
 import com.starrocks.connector.RemoteFileOperations;
 import com.starrocks.connector.exception.StarRocksConnectorException;
 import com.starrocks.connector.hive.PartitionUpdate.UpdateMode;
 import com.starrocks.credential.CloudConfiguration;
+import com.starrocks.qe.ConnectContext;
 import com.starrocks.qe.SessionVariable;
 import com.starrocks.server.GlobalStateMgr;
+import com.starrocks.sql.ast.AddPartitionClause;
+import com.starrocks.sql.ast.AlterClause;
+import com.starrocks.sql.ast.AlterTableStmt;
 import com.starrocks.sql.ast.CreateTableLikeStmt;
 import com.starrocks.sql.ast.CreateTableStmt;
 import com.starrocks.sql.ast.DropTableStmt;
+import com.starrocks.sql.ast.SingleItemListPartitionDesc;
 import com.starrocks.sql.optimizer.OptimizerContext;
 import com.starrocks.sql.optimizer.operator.scalar.ColumnRefOperator;
 import com.starrocks.sql.optimizer.operator.scalar.ScalarOperator;
@@ -47,6 +57,7 @@ import com.starrocks.sql.optimizer.statistics.Statistics;
 import com.starrocks.statistic.StatisticUtils;
 import com.starrocks.thrift.THiveFileInfo;
 import com.starrocks.thrift.TSinkCommitInfo;
+import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.Path;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -159,20 +170,15 @@ public class HiveMetadata implements ConnectorMetadata {
                         hmsTable.getDbName(), hmsTable.getTableName(), hmsTable.getTableLocation()));
             }
         } else {
-            if (!stmt.isForceDrop()) {
-                throw new DdlException(String.format("Table location will be cleared." +
-                        " 'Force' must be set when dropping a hive table." +
-                        " Please execute 'drop table %s.%s.%s force'", stmt.getCatalogName(), dbName, tableName));
-            }
-
             HiveTable hiveTable = (HiveTable) getTable(dbName, tableName);
             if (hiveTable == null && stmt.isSetIfExists()) {
                 LOG.warn("Table {}.{} doesn't exist", dbName, tableName);
                 return;
             }
-
-            if (hiveTable.getHiveTableType() != HiveTable.HiveTableType.MANAGED_TABLE) {
-                throw new StarRocksConnectorException("Only support to drop hive managed table");
+            if (hiveTable.getHiveTableType() == HiveTable.HiveTableType.MANAGED_TABLE && !stmt.isForceDrop()) {
+                throw new DdlException(String.format("Table location will be cleared." +
+                        " 'Force' must be set when dropping a hive table." +
+                        " Please execute 'drop table %s.%s.%s force'", stmt.getCatalogName(), dbName, tableName));
             }
 
             hmsOps.dropTable(dbName, tableName);
@@ -238,6 +244,52 @@ public class HiveMetadata implements ConnectorMetadata {
         }
 
         return fileOps.getRemoteFiles(partitions.build(), useRemoteFileCache);
+    }
+
+    @Override
+    public List<RemoteFileInfo> getRemoteFileInfos(Table table, List<String> partitionNames) {
+        ImmutableList.Builder<Partition> partitions = ImmutableList.builder();
+        Map<String, Partition> existingPartitions = hmsOps.getPartitionByNames(table, partitionNames);
+        partitions.addAll(existingPartitions.values());
+        boolean useRemoteFileCache = true;
+        if (table instanceof HiveTable) {
+            useRemoteFileCache = ((HiveTable) table).isUseMetadataCache();
+        }
+        return fileOps.getRemoteFiles(partitions.build(), useRemoteFileCache);
+    }
+
+    @Override
+    public List<RemoteFileInfo> getRemoteFileInfoForPartitions(Table table, List<String> partitionNames) {
+        ImmutableList.Builder<Partition> partitionBuilder = ImmutableList.builder();
+        Map<String, Partition> existingPartitions = hmsOps.getPartitionByNames(table, partitionNames);
+        partitionBuilder.addAll(existingPartitions.values());
+
+        List<Partition> partitions = partitionBuilder.build();
+        List<Path> paths = Lists.newArrayList();
+        for (Partition partition : partitions) {
+            Path partitionPath = new Path(partition.getFullPath());
+            paths.add(partitionPath);
+        }
+        FileStatus[] fileStatuses = fileOps.getFileStatus(paths.toArray(new Path[0]));
+        List<RemoteFileInfo> remoteFileInfos = Lists.newArrayList();
+        for (int i = 0; i < partitions.size(); i++) {
+            Partition partition = partitions.get(i);
+            FileStatus fileStatus = fileStatuses[i];
+            String locateName = fileStatus.getPath().toUri().getPath();
+            String fileName = PartitionUtil.getSuffixName(paths.get(i).toUri().getPath(), locateName);
+            RemoteFileDesc fileDesc = new RemoteFileDesc(fileName, "", fileStatus.getLen(),
+                    fileStatus.getModificationTime(), ImmutableList.of(),
+                    ImmutableList.of());
+            RemoteFileInfo.Builder builder = RemoteFileInfo.builder()
+                    .setFormat(partition.getInputFormat())
+                    .setFullPath(partition.getFullPath())
+                    .setFiles(List.of(fileDesc).stream()
+                            .map(desc -> desc.setTextFileFormatDesc(partition.getTextFileFormatDesc()))
+                            .map(desc -> desc.setSplittable(partition.isSplittable()))
+                            .collect(Collectors.toList()));
+            remoteFileInfos.add(builder.build());
+        }
+        return remoteFileInfos;
     }
 
     @Override
@@ -373,5 +425,50 @@ public class HiveMetadata implements ConnectorMetadata {
     @Override
     public CloudConfiguration getCloudConfiguration() {
         return hdfsEnvironment.getCloudConfiguration();
+    }
+
+    @Override
+    public void alterTable(AlterTableStmt stmt) throws UserException {
+        // (FIXME) add this api just for tests of external table
+        List<AlterClause> alterClauses = stmt.getOps();
+        for (AlterClause alterClause : alterClauses) {
+            if (alterClause instanceof AddPartitionClause) {
+                addPartition(stmt, alterClause);
+            } else {
+                throw new StarRocksConnectorException("This connector doesn't support alter table type: %s",
+                        alterClause.getOpType());
+            }
+        }
+    }
+
+    private void addPartition(AlterTableStmt stmt, AlterClause alterClause) {
+        HiveTable table = (HiveTable) getTable(stmt.getDbName(), stmt.getTableName());
+        AddPartitionClause addPartitionClause = (AddPartitionClause) alterClause;
+        List<String> partitionColumns = table.getPartitionColumnNames();
+        // now do not support to specify location of hive partition in add partition
+        if (!(addPartitionClause.getPartitionDesc() instanceof SingleItemListPartitionDesc)) {
+            return;
+        }
+        SingleItemListPartitionDesc partitionDesc = (SingleItemListPartitionDesc) addPartitionClause.getPartitionDesc();
+        String tablePath = table.getTableLocation();
+        String partitionString = partitionColumns.get(0) + "=" + partitionDesc.getValues().get(0);
+        String partitionPath = tablePath + "/" + partitionString;
+        HivePartition hivePartition = HivePartition.builder()
+                .setDatabaseName(table.getDbName())
+                .setTableName(table.getTableName())
+                .setColumns(table.getDataColumnNames().stream()
+                        .map(table::getColumn)
+                        .collect(Collectors.toList()))
+                .setValues(partitionDesc.getValues())
+                .setParameters(ImmutableMap.<String, String>builder()
+                        .put("starrocks_version", Version.STARROCKS_VERSION + "-" + Version.STARROCKS_COMMIT_HASH)
+                        .put(STARROCKS_QUERY_ID, ConnectContext.get().getQueryId().toString())
+                        .buildOrThrow())
+                .setStorageFormat(table.getStorageFormat())
+                .setLocation(partitionPath)
+                .build();
+        HivePartitionWithStats partitionWithStats =
+                new HivePartitionWithStats(partitionString, hivePartition, HivePartitionStats.empty());
+        hmsOps.addPartitions(table.getDbName(), table.getTableName(), Lists.newArrayList(partitionWithStats));
     }
 }
